Moved slot handling code to its own file
