+++ src/cap32.h
@@ -471,23 +471,6 @@ void audio_pause (void);
 void audio_resume (void);
 int video_init (void);
 void video_shutdown (void);
-int snapshot_load (FILE *pfile);
-int snapshot_load (const std::string &filename);
-int snapshot_save (const std::string &filename);
-int dsk_load (FILE *pfile, t_drive *drive);
-int dsk_load (const std::string &filename, t_drive *drive);
-int dsk_save (const std::string &filename, t_drive *drive);
-void dsk_eject (t_drive *drive);
-int dsk_format (t_drive *drive, int iFormat);
-int tape_insert (FILE *pfile);
-int tape_insert (const std::string &filename);
-int tape_insert_cdt (FILE *pfile);
-int tape_insert_voc (FILE *pfile);
-void tape_eject (void);
-
-// Smart load: support loading DSK, SNA, CDT, VOC, CPR or a zip containing one of these.
-// drive must be DSK_A or DSK_B for DSK, OTHER otherwise.
-int file_load(const std::string& filepath, const DRIVE drive);
 
 // Return the path to the best (i.e: most specific) configuration file.
 // Priority order is:
@@ -495,13 +478,8 @@ int file_load(const std::string& filepath, const DRIVE drive);
 //  - $HOME/.cap32.cfg
 //  - /etc/cap32.cfg
 std::string getConfigurationFilename(bool forWrite = false);
-t_disk_format parseDiskFormat(const std::string& format);
-std::string serializeDiskFormat(const t_disk_format& format);
 void loadConfiguration (t_CPC &CPC, const std::string& configFilename);
 void saveConfiguration (t_CPC &CPC, const std::string& configFilename);
-
-// Retrieve files that are passed as argument and update CPC fields so that they will be loaded properly
-void fillSlots (std::vector<std::string> slot_list, t_CPC& CPC);
 
 int cap32_main(int argc, char **argv);
